divided whole program into multiple files for modular design
